DDP-7306. Question extended with scale (type fix)

+++ pepper-apis/dss-core/src/main/java/org/broadinstitute/ddp/db/dao/JdbiDecimalQuestion.java
@@ -4,13 +4,11 @@ import org.jdbi.v3.sqlobject.SqlObject;
 import org.jdbi.v3.sqlobject.customizer.Bind;
 import org.jdbi.v3.sqlobject.statement.SqlUpdate;
 
-import java.math.BigInteger;
-
 public interface JdbiDecimalQuestion extends SqlObject {
 
     @SqlUpdate("insert into decimal_question (question_id, placeholder_template_id, scale)"
             + " values (:questionId, :placeholderTemplateId, :scale)")
     int insert(@Bind("questionId") long questionId,
                @Bind("placeholderTemplateId") Long placeholderTemplateId,
-               @Bind("scale") BigInteger scale);
+               @Bind("scale") Integer scale);
 }
