+++ pepper-apis/dss-core/src/main/java/org/broadinstitute/ddp/db/dto/DecimalQuestionDto.java
@@ -5,17 +5,16 @@ import org.jdbi.v3.core.mapper.reflect.ColumnName;
 import org.jdbi.v3.core.mapper.reflect.JdbiConstructor;
 
 import java.io.Serializable;
-import java.math.BigInteger;
 import java.util.Set;
 
 public final class DecimalQuestionDto extends QuestionDto implements Serializable {
     private final Long placeholderTemplateId;
-    private final BigInteger scale;
+    private final Integer scale;
 
     @JdbiConstructor
     public DecimalQuestionDto(@Nested QuestionDto questionDto,
                               @ColumnName("placeholder_template_id") Long placeholderTemplateId,
-                              @ColumnName("scale") BigInteger scale) {
+                              @ColumnName("scale") Integer scale) {
         super(questionDto);
         this.placeholderTemplateId = placeholderTemplateId;
         this.scale = scale;
@@ -25,7 +24,7 @@ public final class DecimalQuestionDto extends QuestionDto implements Serializabl
         return placeholderTemplateId;
     }
 
-    public BigInteger getScale() {
+    public Integer getScale() {
         return scale;
     }
